IC-535 Clean up of dtos
